Rename model.h -> Library.h
